+++ forester/models/__init__.py
@@ -7,6 +7,8 @@ from .blob import Blob
 from .tree import Tree, TreeEntry
 from .commit import Commit
 from .mesh import Mesh
+from .mesh_diff import MeshDiff, GeometryDiff, MaterialDiff, DiffStatistics
+from .texture import Texture
 
 __all__ = [
     "Blob",
@@ -14,6 +16,11 @@ __all__ = [
     "TreeEntry",
     "Commit",
     "Mesh",
+    "MeshDiff",
+    "GeometryDiff",
+    "MaterialDiff",
+    "DiffStatistics",
+    "Texture",
 ]